refactor: implementation of query package and utils with temporary main
