+++ src/query/callWorker.go
@@ -1,32 +1,33 @@
-package main
+package query
 
 import (
-  "time"
+	"bruteforce/src/matching"
+	"bruteforce/src/utils"
+	"time"
 )
 
-type forceData struct {
-	worker int
-	wordList string
-	url string
+type ForceData struct {
+	Worker   int
+	WordList string
+	Url      string
 }
 
-func executeQueryFromFile(data forceData, currentPath chan string) {
-  for taskData := range currentPath{
-    queryExecute(data, taskData, "POST")
-    }
+func executeQueryFromFile(data ForceData, currentPath chan string) {
+	for taskData := range currentPath {
+		QueryExecute(data, taskData, "POST")
+	}
 }
 
-func mainRequest(data forceData) {
-  channel := make(chan string)
-  wordArray := GetFileContent("../wordList/rootList")
+func MainRequest(data ForceData, criteria matcher.MatchCriteria) {
+	channel := make(chan string)
+	wordArray := utils.GetFileContent(data.WordList)
 
-  for i := 0 ;i < data.worker; i++ {
+	for i := 0; i < data.Worker; i++ {
 		go executeQueryFromFile(data, channel)
 	}
-  for i := 0; i < len(wordArray); i++ {
-    channel <- wordArray[i]
-  }
-  time.Sleep(1 * time.Second)
-  close(channel)
+	for i := 0; i < len(wordArray); i++ {
+		channel <- wordArray[i]
+	}
+	time.Sleep(1 * time.Second)
+	close(channel)
 }
-
