+++ src/query/queryExecute.go
@@ -1,35 +1,32 @@
-package main
+package query
 
 import (
 	"fmt"
+	"io"
 	"log"
-	"io/ioutil"
 	"net/http"
 )
 
-func queryExecute(data forceData, path string, method string) {
-
+func QueryExecute(data ForceData, path string, method string) {
 	client := &http.Client{}
-	req, err := http.NewRequest(method, data.url + path, nil)
+	req, err := http.NewRequest(method, data.Url+path, nil)
 	if err != nil {
-	  log.Fatal(err)
+		log.Fatal(err)
 	}
 
 	q := req.URL.Query()
-
 	req.URL.RawQuery = q.Encode()
 
 	resp, err := client.Do(req)
 	if err != nil {
-	  fmt.Println(err)
+		fmt.Println(err)
 	}
-
 	defer resp.Body.Close()
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-	  log.Fatal(err)
+		log.Fatal(err)
 	}
 
 	fmt.Println(string(body))
-  }
+}
