+++ src/utils/getFile.go
@@ -1,18 +1,17 @@
-package main
+package utils
 
 import (
-    "os"
-    "fmt"
-    "log"
+	"log"
+	"os"
 	"strings"
 )
 
-func GetFileContent(filePath string) []string{
-    body, err := os.ReadFile(filePath)
-    if err != nil {
-        log.Fatalf("unable to read file: %v", err)
-    }
+func GetFileContent(filePath string) []string {
+	body, err := os.ReadFile(filePath)
+	if err != nil {
+		log.Fatalf("Unable to read file: %v", err)
+	}
+
 	dataTab := strings.Split(string(body), "\n")
-    fmt.Println(dataTab[0])
 	return dataTab
 }
